refactor(proxy): split multipart.go into focused files (#770)
internal/proxy/multipart.go had grown to 816 lines covering create,
upload, complete, abort, cleanup, encryption, locking, and assembly
streaming in one place. Splits it into focused siblings without changing
behaviour:

- multipart.go         type, ctor, CreateMultipartUpload, UploadPart,
                       ListMultipartUploads, GetParts (entry points)
- multipart_encrypt.go DEK unwrap/forget, prepareUploadPartBody, new
                       shared encryptWithUploadDEK helper
- multipart_lock.go    uploadIDLockNamespace + uploadIDLockID
- multipart_parts.go   multipartPartKey, fetchScopedUpload,
                       validateMultipartScope, collectRequestedParts,
                       formatPartNumbers
- multipart_complete.go CompleteMultipartUpload, locked path,
                       cleanupCompletedUpload, sumPlaintextSize,
                       buildAssembledUpload, streamPartsThroughPipe,
                       streamOnePart
- multipart_cleanup.go AbortMultipartUpload, abortByMultipartRow,
                       CleanupStaleMultipartUploads,
                       AbortMultipartUploadsOnBackend

Additional cleanup: extracts the duplicated encryption setup from
prepareUploadPartBody and buildAssembledUpload into a single
encryptWithUploadDEK helper. Both callers now do their gate
(encryptor != nil / mu.Encrypted), call the shared helper, and wrap
the error with call-site-specific context. The unwrap + EncryptWithDEK
+ EncryptionMeta-pack pattern lives in one place.

manager_multipart_test.go is unchanged - tests stay in package proxy
so symbol moves within the package do not affect them.
